publishModal: move modalItem to module
