+++ main.ts
@@ -3,7 +3,7 @@ import Publisher from "./src/publisher/Publisher";
 import DigitalGardenSettings from "./src/models/settings";
 import { PublishStatusBar } from "./src/ui/PublishStatusBar";
 import { seedling } from "src/ui/suggest/constants";
-import { PublishModal } from "src/ui/PublishModal";
+import { PublishModal } from "src/ui/PublishModal/PublishModal";
 import PublishStatusManager from "src/publisher/PublishStatusManager";
 import ObsidianFrontMatterEngine from "src/publisher/ObsidianFrontMatterEngine";
 import DigitalGardenSiteManager from "src/publisher/DigitalGardenSiteManager";
